Добавлена послойная архитектура как в CleanArch. Настроены зависимости между проектами.
