+++ CleanArchEmpty.sln
@@ -1,0 +1,85 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.10.34607.79
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "WebApi", "WebApi", "{3ADE0793-A368-4745-9A0F-93EDA9B43262}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Domain", "Domain", "{C7B923AB-9DE5-469D-A0F7-86151DF421D0}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Infrastructure", "Infrastructure", "{7B014471-0A88-4A8C-97FF-BF6AAAA1AF84}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Services", "Services", "{EC83CFB6-0AE4-453D-A39C-4CD07F4F3A51}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain.Entities", "Domain\Domain.Entities\Domain.Entities.csproj", "{82A5BE82-C1A5-4F51-8ED1-E2CE1A099C96}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastructure.EntityFramework", "Infrastructure\Infrastructure.EntityFramework\Infrastructure.EntityFramework.csproj", "{FD1D3B84-B31E-40D0-B8EA-FCA78FC87F2A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastructure.Repositories.Implementations", "Infrastructure\Infrastructure.Repositories.Implementations\Infrastructure.Repositories.Implementations.csproj", "{ACC667ED-DC39-4A66-AA10-68E9085741BF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Services.Abstractions", "Services\Services.Abstractions\Services.Abstractions.csproj", "{BC7158A0-76EA-4A10-BF7E-04033660B7B0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Services.Contracts", "Services\Services.Contracts\Services.Contracts.csproj", "{54908A0A-725E-455F-8A79-9C0C21F25BAF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Services.Implementations", "Services\Services.Implementations\Services.Implementations.csproj", "{D52BE77D-E4C1-4B4E-9C82-093B5F84EBA9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Services.Repositories.Abstractions", "Services\Services.Repositories.Abstractions\Services.Repositories.Abstractions.csproj", "{87E7F64A-1A41-4F9C-82F4-53C633628165}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApi", "WebApi\WebApi\WebApi.csproj", "{E4E78E4B-DD21-4146-89BA-20E22820BA7C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{82A5BE82-C1A5-4F51-8ED1-E2CE1A099C96}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{82A5BE82-C1A5-4F51-8ED1-E2CE1A099C96}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{82A5BE82-C1A5-4F51-8ED1-E2CE1A099C96}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{82A5BE82-C1A5-4F51-8ED1-E2CE1A099C96}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FD1D3B84-B31E-40D0-B8EA-FCA78FC87F2A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FD1D3B84-B31E-40D0-B8EA-FCA78FC87F2A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FD1D3B84-B31E-40D0-B8EA-FCA78FC87F2A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FD1D3B84-B31E-40D0-B8EA-FCA78FC87F2A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{ACC667ED-DC39-4A66-AA10-68E9085741BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ACC667ED-DC39-4A66-AA10-68E9085741BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ACC667ED-DC39-4A66-AA10-68E9085741BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ACC667ED-DC39-4A66-AA10-68E9085741BF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BC7158A0-76EA-4A10-BF7E-04033660B7B0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BC7158A0-76EA-4A10-BF7E-04033660B7B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BC7158A0-76EA-4A10-BF7E-04033660B7B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BC7158A0-76EA-4A10-BF7E-04033660B7B0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{54908A0A-725E-455F-8A79-9C0C21F25BAF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{54908A0A-725E-455F-8A79-9C0C21F25BAF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{54908A0A-725E-455F-8A79-9C0C21F25BAF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{54908A0A-725E-455F-8A79-9C0C21F25BAF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D52BE77D-E4C1-4B4E-9C82-093B5F84EBA9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D52BE77D-E4C1-4B4E-9C82-093B5F84EBA9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D52BE77D-E4C1-4B4E-9C82-093B5F84EBA9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D52BE77D-E4C1-4B4E-9C82-093B5F84EBA9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{87E7F64A-1A41-4F9C-82F4-53C633628165}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{87E7F64A-1A41-4F9C-82F4-53C633628165}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{87E7F64A-1A41-4F9C-82F4-53C633628165}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{87E7F64A-1A41-4F9C-82F4-53C633628165}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E4E78E4B-DD21-4146-89BA-20E22820BA7C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E4E78E4B-DD21-4146-89BA-20E22820BA7C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E4E78E4B-DD21-4146-89BA-20E22820BA7C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E4E78E4B-DD21-4146-89BA-20E22820BA7C}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{82A5BE82-C1A5-4F51-8ED1-E2CE1A099C96} = {C7B923AB-9DE5-469D-A0F7-86151DF421D0}
+		{FD1D3B84-B31E-40D0-B8EA-FCA78FC87F2A} = {7B014471-0A88-4A8C-97FF-BF6AAAA1AF84}
+		{ACC667ED-DC39-4A66-AA10-68E9085741BF} = {7B014471-0A88-4A8C-97FF-BF6AAAA1AF84}
+		{BC7158A0-76EA-4A10-BF7E-04033660B7B0} = {EC83CFB6-0AE4-453D-A39C-4CD07F4F3A51}
+		{54908A0A-725E-455F-8A79-9C0C21F25BAF} = {EC83CFB6-0AE4-453D-A39C-4CD07F4F3A51}
+		{D52BE77D-E4C1-4B4E-9C82-093B5F84EBA9} = {EC83CFB6-0AE4-453D-A39C-4CD07F4F3A51}
+		{87E7F64A-1A41-4F9C-82F4-53C633628165} = {EC83CFB6-0AE4-453D-A39C-4CD07F4F3A51}
+		{E4E78E4B-DD21-4146-89BA-20E22820BA7C} = {3ADE0793-A368-4745-9A0F-93EDA9B43262}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {31E829B5-E092-47E2-B0E7-C2667F644049}
+	EndGlobalSection
+EndGlobal
