+++ Domain/Domain.Entities/BaseEntity.cs
@@ -1,0 +1,7 @@
+﻿namespace Domain.Entities
+{
+    public class BaseEntity
+    {
+        public Guid Id { get; set; }
+    }
+}
