+++ Domain/Domain.Entities/Employee.cs
@@ -1,0 +1,17 @@
+﻿
+namespace Domain.Entities
+{
+    public class Employee : BaseEntity
+    {
+        public string FirstName { get; set; }
+        public string LastName { get; set; }
+
+        public string FullName => $"{FirstName} {LastName}";
+
+        public string Email { get; set; }
+
+        public List<Role> Roles { get; set; }
+
+        public int AppliedPromocodesCount { get; set; }
+    }
+}
