+++ Domain/Domain.Entities/Role.cs
@@ -1,0 +1,10 @@
+﻿
+namespace Domain.Entities
+{
+    public class Role : BaseEntity
+    {
+        public string Name { get; set; }
+
+        public string Description { get; set; }
+    }
+}
