+++ WebApi/WebApi/Models/EmployeeResponse.cs
@@ -1,0 +1,15 @@
+﻿
+namespace Otus.Teaching.PromoCodeFactory.WebHost.Models
+{
+    public class EmployeeResponse
+    {
+        public Guid Id { get; set; }
+        public string FullName { get; set; }
+
+        public string Email { get; set; }
+
+        public List<RoleItemResponse> Roles { get; set; }
+
+        public int AppliedPromocodesCount { get; set; }
+    }
+}
