+++ WebApi/WebApi/Models/InputEmployeeModel.cs
@@ -1,0 +1,11 @@
+﻿
+namespace Otus.Teaching.PromoCodeFactory.WebHost.Models
+{
+    public class InputEmployeeModel
+    {
+        public string FirstName { get; set; }
+        public string LastName { get; set; }
+        public string Email { get; set; }
+        public int AppliedPromocodesCount { get; set; }
+    }
+}
