+++ WebApi/WebApi/Models/RoleItemResponse.cs
@@ -1,0 +1,11 @@
+﻿
+namespace Otus.Teaching.PromoCodeFactory.WebHost.Models
+{
+    public class RoleItemResponse
+    {
+        public Guid Id { get; set; }
+        public string Name { get; set; }
+
+        public string Description { get; set; }
+    }
+}
