+++ WebApi/WebApi/Startup.cs
@@ -1,0 +1,51 @@
+﻿
+using Domain.Entities;
+using Infrastructure.EntityFramework;
+using Infrastructure.Repositories.Implementations;
+using Otus.Teaching.PromoCodeFactory.WebHost.Models;
+using Services.Repositories.Abstractions;
+
+namespace Otus.Teaching.PromoCodeFactory.WebHost
+{
+    public class Startup
+    {
+        // This method gets called by the runtime. Use this method to add services to the container.
+        // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
+        public void ConfigureServices(IServiceCollection services)
+        {
+            services.AddControllers();
+            services.AddSingleton(typeof(IRepository<Employee>), (x) =>
+                new InMemoryRepository<Employee>(FakeDataFactory.Employees));
+            services.AddSingleton(typeof(IRepository<Role>), (x) =>
+                new InMemoryRepository<Role>(FakeDataFactory.Roles));
+            services.AddSingleton<EmployeeHelper>();
+
+            services.AddEndpointsApiExplorer();
+            services.AddSwaggerGen();
+        }
+
+        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+        {
+            if (env.IsDevelopment())
+            {
+                app.UseDeveloperExceptionPage();
+                app.UseSwagger();
+                app.UseSwaggerUI();
+            }
+            else
+            {
+                app.UseHsts();
+            }
+
+            app.UseHttpsRedirection();
+
+            app.UseRouting();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
+        }
+    }
+}
