--- src/Otus.Teaching.PromoCodeFactory.Core/Abstractions/Repositories/IRepository.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using Otus.Teaching.PromoCodeFactory.Core.Domain;
-
-namespace Otus.Teaching.PromoCodeFactory.Core.Abstractions.Repositories
-{
-    public interface IRepository<T>
-        where T: BaseEntity
-    {
-        Task<IEnumerable<T>> GetAllAsync();
-        
-        Task<T> GetByIdAsync(Guid id);
-
-        Task<T> CreateAsync(T entity);
-
-        Task UpdateAsync(T oldEntity, T newEntity);
-
-        Task DeleteAsync(T entity);
-    }
-}
