--- src/Otus.Teaching.PromoCodeFactory.Core/Domain/Administration/Role.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace Otus.Teaching.PromoCodeFactory.Core.Domain.Administration
-{
-    public class Role
-        : BaseEntity
-    {
-        public string Name { get; set; }
-
-        public string Description { get; set; }
-    }
-}
