--- src/Otus.Teaching.PromoCodeFactory.Core/Domain/BaseEntity.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Otus.Teaching.PromoCodeFactory.Core.Domain
-{
-    public class BaseEntity
-    {
-        public Guid Id { get; set; }
-    }
-}
