--- src/Otus.Teaching.PromoCodeFactory.WebHost/Models/EmployeeResponse.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using Otus.Teaching.PromoCodeFactory.Core.Domain.Administration;
-
-namespace Otus.Teaching.PromoCodeFactory.WebHost.Models
-{
-    public class EmployeeResponse
-    {
-        public Guid Id { get; set; }
-        public string FullName { get; set; }
-
-        public string Email { get; set; }
-
-        public List<RoleItemResponse> Roles { get; set; }
-
-        public int AppliedPromocodesCount { get; set; }
-    }
-}
