--- src/Otus.Teaching.PromoCodeFactory.WebHost/Models/InputEmployeeModel.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace Otus.Teaching.PromoCodeFactory.WebHost.Models
-{
-    public class InputEmployeeModel
-    {
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Email { get; set; }
-        public int AppliedPromocodesCount { get; set; }
-    }
-}
