--- src/Otus.Teaching.PromoCodeFactory.WebHost/Models/RoleItemResponse.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace Otus.Teaching.PromoCodeFactory.WebHost.Models
-{
-    public class RoleItemResponse
-    {
-        public Guid Id { get; set; }
-        public string Name { get; set; }
-
-        public string Description { get; set; }
-    }
-}
