--- src/Otus.Teaching.PromoCodeFactory.WebHost/Properties/launchSettings.json
@@ -1,13 +1,0 @@
-﻿{
-  "profiles": {
-    "Otus.Teaching.PromocodeFactory.WebHost": {
-      "commandName": "Project",
-      "launchUrl": "swagger/index.html",
-      "launchBrowser": true,
-      "applicationUrl": "https://localhost:5001;http://localhost:5000",
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
-      }
-    }
-  }
-}
