--- src/Otus.Teaching.PromoCodeFactory.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Otus.Teaching.PromoCodeFactory.WebHost", "Otus.Teaching.PromoCodeFactory.WebHost\Otus.Teaching.PromoCodeFactory.WebHost.csproj", "{C086AC72-ADC7-432C-99D8-A8D60A0B44CA}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Otus.Teaching.PromoCodeFactory.DataAccess", "Otus.Teaching.PromoCodeFactory.DataAccess\Otus.Teaching.PromoCodeFactory.DataAccess.csproj", "{D7E675A3-9872-47D6-B906-797B7B866534}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Otus.Teaching.PromoCodeFactory.Core", "Otus.Teaching.PromoCodeFactory.Core\Otus.Teaching.PromoCodeFactory.Core.csproj", "{FA64A24E-D404-4E7D-89A8-B87E134F697F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C086AC72-ADC7-432C-99D8-A8D60A0B44CA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C086AC72-ADC7-432C-99D8-A8D60A0B44CA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C086AC72-ADC7-432C-99D8-A8D60A0B44CA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C086AC72-ADC7-432C-99D8-A8D60A0B44CA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D7E675A3-9872-47D6-B906-797B7B866534}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D7E675A3-9872-47D6-B906-797B7B866534}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D7E675A3-9872-47D6-B906-797B7B866534}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D7E675A3-9872-47D6-B906-797B7B866534}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FA64A24E-D404-4E7D-89A8-B87E134F697F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FA64A24E-D404-4E7D-89A8-B87E134F697F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FA64A24E-D404-4E7D-89A8-B87E134F697F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FA64A24E-D404-4E7D-89A8-B87E134F697F}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
